requirements.txt
APScheduler==3.9.1.post1
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
commonmark==0.9.1
Django==4.1.5
django-cors-headers==3.13.0
django-crontab==0.7.1
django-sslserver==0.22
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.2
pycodestyle==2.10.0
Pygments==2.14.0
pyotp==2.8.0
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
six==1.16.0
smartapi-python==1.3.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
websocket-client==1.4.2
